pyproject.toml
[tool.poetry]
name = "participle"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pillow = "^10.4.0"
pytesseract = "^0.3.13"
thefuzz = "^0.22.1"
gspread = "^6.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

